pyproject.toml
[project]
name = "jarvis-mac"
version = "0.1.0"
description = "Autonomous Mac-native development partner powered by Claude Agent SDK and Apple Containers"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "click>=8.1",
    "rich>=13.0",
    "mcp>=1.0",
    "httpx>=0.27",
    "websockets>=12.0",
]

[project.optional-dependencies]
slack = ["slack-bolt>=1.18", "slack-sdk>=3.27"]
voice = ["pyobjc-framework-AVFoundation>=10.0"]
all = ["slack-bolt>=1.18", "slack-sdk>=3.27", "pyobjc-framework-AVFoundation>=10.0", "websockets>=12.0"]

[project.scripts]
jarvis = "jarvis.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
